LayoutTests/fast/events/scroll-after-click-on-tab-index-expected.txt

ALERT: Scroll position 0
Click the text below and then try to scroll.
PASS Scroll position is more than 0
PASS successfullyParsed is true

TEST COMPLETE










My span
